requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
pandas==2.2.3
scikit-learn==1.6.0
dotenv==0.9.9
apache-airflow-providers-http==5.3.0
apache-airflow-providers-postgres==6.2.0
mlflow==2.22.0
apache-airflow-providers-amazon==9.8.0
boto3>=1.37.0